requirements.txt
streamlit
pandas
numpy
joblib
requests
psycopg2-binary
plotly
scikit-learn